software/requirements.txt
numpy
pyserial
PyQt5
PyOpenGL
pyqtgraph
